MP2-750 - adds office and administration flat rate on direct costs in the budget options - updates budget costs calculator to consider office and administration flat rate on direct costs in it's calculations

+++ jems-server/src/main/kotlin/io/cloudflight/jems/server/project/repository/partner/budget/ProjectPartnerBudgetOptionsPersistenceProvider.kt
@@ -1,6 +1,5 @@
 package io.cloudflight.jems.server.project.repository.partner.budget
 
-import io.cloudflight.jems.server.project.repository.partner.budget.ProjectPartnerBudgetStaffCostRepository
 import io.cloudflight.jems.server.project.service.partner.budget.ProjectPartnerBudgetOptionsPersistence
 import io.cloudflight.jems.server.project.service.partner.model.ProjectPartnerBudgetOptions
 import org.springframework.stereotype.Repository
@@ -27,24 +26,14 @@ class ProjectPartnerBudgetOptionsPersistenceProvider(
 
     @Transactional
     override fun updateBudgetOptions(partnerId: Long, options: ProjectPartnerBudgetOptions) {
-        budgetOptionsRepository.findById(partnerId).ifPresentOrElse(
-            {
-                it.officeAndAdministrationOnStaffCostsFlatRate = options.officeAndAdministrationOnStaffCostsFlatRate
-                it.travelAndAccommodationOnStaffCostsFlatRate = options.travelAndAccommodationOnStaffCostsFlatRate
-                it.staffCostsFlatRate = options.staffCostsFlatRate
-                it.otherCostsOnStaffCostsFlatRate = options.otherCostsOnStaffCostsFlatRate
-            },
-            {
-                budgetOptionsRepository.save(options.toProjectPartnerBudgetOptionsEntity())
-            }
-        )
+        budgetOptionsRepository.save(options.toProjectPartnerBudgetOptionsEntity())
     }
 
     @Transactional
-    override fun deleteBudgetOptions(partnerId: Long) =
-        budgetOptionsRepository.findById(partnerId).ifPresent {
+    override fun deleteBudgetOptions(partnerId: Long) {
+        if (budgetOptionsRepository.existsById(partnerId))
             budgetOptionsRepository.deleteById(partnerId)
-        }
+    }
 
     @Transactional
     override fun deleteStaffCosts(partnerId: Long) =
